Moves helper methods to the test project
