+++ CasualGodComplex.Test/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="ColorSharp" version="0.11.0.32448" targetFramework="net45" userInstalled="true" />
+  <package id="System.Numerics.Vectors" version="4.0.0" targetFramework="net45" userInstalled="true" />
+</packages>
